use query in commands

+++ my_notes/notes/commands.py
@@ -1,6 +1,5 @@
-from .models import Note
 from .serializers import NoteSerializer
-
+from .queries import get_note as get_note_query
 
 def create_note(data):
     try:
@@ -12,8 +11,12 @@ def create_note(data):
     except Exception as e:
         return str(e)
 
-def update_note(note: Note, data):
+def update_note(pk, data):
     try:
+        note = get_note_query(pk)
+        if not note:
+            return 'Заметка с id = {} не найдена'.format(pk)
+
         srl = NoteSerializer(note, data=data, partial=True)
         if not srl.is_valid():
             return srl.errors
@@ -22,8 +25,12 @@ def update_note(note: Note, data):
     except Exception as e:
         return str(e)
 
-def delete_note(note: Note):
+def delete_note(pk):
     try:
+        note = get_note_query(pk)
+        if not note:
+            return 'Заметка с id = {} не найдена'.format(pk)
+        
         note.delete()
     except Exception as e:
         return str(e)
